+++ .gitignore
@@ -8,6 +8,7 @@
 
 # testing
 /coverage
+/analyze
 
 # next.js
 /.next/
