+++ next.config.mjs
@@ -1,7 +1,7 @@
 import withBundleAnalyzer from '@next/bundle-analyzer';
 
 const bundleAnalyzer = withBundleAnalyzer({
-    enabled: process.env.ANALYZE === 'false',
+    enabled: process.env.ANALYZE === 'true',
 })
